dev-lang/rust: Version bump to 1.41.1

+++ dev-lang/rust/rust-1.41.1.ebuild
@@ -1,30 +1,34 @@
-# Copyright 1999-2019 Gentoo Authors
+# Copyright 1999-2020 Gentoo Authors
 # Distributed under the terms of the GNU General Public License v2
 
 EAPI=7
 
-PYTHON_COMPAT=( python2_7 python3_{5,6,7} pypy )
+PYTHON_COMPAT=( python3_{6,7} )
 
-inherit check-reqs estack flag-o-matic llvm multilib-build multiprocessing python-any-r1 rust-toolchain toolchain-funcs
+inherit bash-completion-r1 check-reqs estack flag-o-matic llvm multilib-build multiprocessing python-any-r1 rust-toolchain toolchain-funcs
 
 ABI_VER="$(ver_cut 1-2)"
 SLOT="stable/${ABI_VER}"
 MY_P="rustc-${PV}"
 SRC="${MY_P}-src.tar.xz"
-KEYWORDS="~amd64 ~arm ~arm64 ~ppc ~ppc64 ~x86"
+KEYWORDS="amd64 ~arm ~arm64 ~ppc ppc64 x86"
 
 RUST_STAGE0_VERSION="1.$(($(ver_cut 2) - 1)).0"
+RUST_TOOLCHAIN_BASEURL="https://portage.smaeul.xyz/distfiles/"
 
 DESCRIPTION="Systems programming language from Mozilla"
 HOMEPAGE="https://www.rust-lang.org/"
 
-SRC_URI="https://static.rust-lang.org/dist/${SRC} -> rustc-${PV}-src.tar.xz
-	amd64? ( https://portage.smaeul.xyz/distfiles/rust-${RUST_STAGE0_VERSION}-x86_64-gentoo-linux-musl.tar.xz )
-	arm? ( https://portage.smaeul.xyz/distfiles/rust-${RUST_STAGE0_VERSION}-armv7a-unknown-linux-musleabihf.tar.xz )
-	arm64? ( https://portage.smaeul.xyz/distfiles/rust-${RUST_STAGE0_VERSION}-aarch64-gentoo-linux-musl.tar.xz )
-	ppc? ( https://portage.smaeul.xyz/distfiles/rust-${RUST_STAGE0_VERSION}-powerpc-gentoo-linux-musl.tar.xz )
-	ppc64? ( https://portage.smaeul.xyz/distfiles/rust-${RUST_STAGE0_VERSION}-powerpc64-gentoo-linux-musl.tar.xz )
-	x86? ( https://portage.smaeul.xyz/distfiles/rust-${RUST_STAGE0_VERSION}-i686-gentoo-linux-musl.tar.xz )
+SRC_URI="
+	https://static.rust-lang.org/dist/${SRC} -> rustc-${PV}-src.tar.xz
+	!system-bootstrap? (
+		amd64? ( $(rust_arch_uri x86_64-gentoo-linux-musl        rust-${RUST_STAGE0_VERSION} ) )
+		arm?   ( $(rust_arch_uri armv7a-unknown-linux-musleabihf rust-${RUST_STAGE0_VERSION} ) )
+		arm64? ( $(rust_arch_uri aarch64-gentoo-linux-musl       rust-${RUST_STAGE0_VERSION} ) )
+		ppc?   ( $(rust_arch_uri powerpc-gentoo-linux-musl       rust-${RUST_STAGE0_VERSION} ) )
+		ppc64? ( $(rust_arch_uri powerpc64-gentoo-linux-musl     rust-${RUST_STAGE0_VERSION} ) )
+		x86?   ( $(rust_arch_uri i686-gentoo-linux-musl          rust-${RUST_STAGE0_VERSION} ) )
+	)
 "
 
 ALL_LLVM_TARGETS=( AArch64 AMDGPU ARM BPF Hexagon Lanai Mips MSP430
@@ -34,7 +38,7 @@ LLVM_TARGET_USEDEPS=${ALL_LLVM_TARGETS[@]/%/?}
 
 LICENSE="|| ( MIT Apache-2.0 ) BSD-1 BSD-2 BSD-4 UoI-NCSA"
 
-IUSE="clippy cpu_flags_arm_neon cpu_flags_arm_thumb2 cpu_flags_x86_sse2 debug doc libressl miri rls rustfmt system-llvm wasm ${ALL_LLVM_TARGETS[*]}"
+IUSE="clippy cpu_flags_arm_neon cpu_flags_arm_thumb2 cpu_flags_x86_sse2 debug doc libressl miri nightly parallel-compiler rls rustfmt system-bootstrap system-llvm wasm ${ALL_LLVM_TARGETS[*]}"
 
 # Please keep the LLVM dependency block separate. Since LLVM is slotted,
 # we need to *really* make sure we're not pulling more than one slot
@@ -42,28 +46,30 @@ IUSE="clippy cpu_flags_arm_neon cpu_flags_arm_thumb2 cpu_flags_x86_sse2 debug do
 
 # How to use it:
 # 1. List all the working slots (with min versions) in ||, newest first.
-# 2. Update the := to specify *max* version, e.g. < 9.
-# 3. Specify LLVM_MAX_SLOT, e.g. 8.
+# 2. Update the := to specify *max* version, e.g. < 11.
+# 3. Specify LLVM_MAX_SLOT, e.g. 10.
 LLVM_DEPEND="
 	|| (
-		>=sys-devel/llvm-9:=[${LLVM_TARGET_USEDEPS// /,}]
-		wasm? ( =sys-devel/lld-9* )
+		sys-devel/llvm:9[${LLVM_TARGET_USEDEPS// /,}]
+		sys-devel/llvm:10[${LLVM_TARGET_USEDEPS// /,}]
 	)
-	<sys-devel/llvm-10:=
+	<sys-devel/llvm-11:=
 "
-LLVM_MAX_SLOT=9
+LLVM_MAX_SLOT=10
+
+BOOTSTRAP_DEPEND="|| ( >=dev-lang/rust-1.$(($(ver_cut 2) - 1)).0-r1 >=dev-lang/rust-bin-1.$(($(ver_cut 2) - 1)) )"
 
 COMMON_DEPEND="
-	!libressl? ( dev-libs/openssl:0= )
-	libressl? ( dev-libs/libressl:0= )
-	dev-libs/libgit2:=
 	net-libs/http-parser:=
 	net-libs/libssh2:=
 	net-misc/curl:=[ssl]
 	sys-libs/libssp_nonshared:=[${MULTILIB_USEDEP}]
 	sys-libs/zlib:=
+	!libressl? ( dev-libs/openssl:0= )
+	libressl? ( dev-libs/libressl:0= )
 	system-llvm? (
 		${LLVM_DEPEND}
+		wasm? ( >=sys-devel/lld-9 )
 	)
 "
 
@@ -73,6 +79,7 @@ DEPEND="${COMMON_DEPEND}
 		>=sys-devel/gcc-4.7
 		>=sys-devel/clang-3.5
 	)
+	system-bootstrap? ( ${BOOTSTRAP_DEPEND}	)
 	!system-llvm? (
 		dev-util/cmake
 		dev-util/ninja
@@ -81,43 +88,48 @@ DEPEND="${COMMON_DEPEND}
 
 RDEPEND="${COMMON_DEPEND}
 	>=app-eselect/eselect-rust-20190311
-	!dev-util/cargo
-	rustfmt? ( !dev-util/rustfmt )
 "
 
 REQUIRED_USE="|| ( ${ALL_LLVM_TARGETS[*]} )
+	parallel-compiler? ( nightly )
 	wasm? ( llvm_targets_WebAssembly )
 	x86? ( cpu_flags_x86_sse2 )
 "
 
+QA_FLAGS_IGNORED="
+	usr/bin/*-${PV}
+	usr/lib*/lib*.so
+	usr/lib/rustlib/*/codegen-backends/librustc_codegen_llvm-llvm.so
+	usr/lib/rustlib/*/lib/lib*.so
+"
+
+QA_SONAME="usr/lib.*/librustc_macros.*.so"
+
 PATCHES=(
-	"${FILESDIR}/0001-Don-t-pass-CFLAGS-to-the-C-compiler.patch"
-	"${FILESDIR}/0002-Fix-LLVM-build.patch"
-	"${FILESDIR}/0003-Allow-rustdoc-to-work-when-cross-compiling-on-musl.patch"
-	"${FILESDIR}/0004-Require-static-native-libraries-when-linking-static-.patch"
-	"${FILESDIR}/0005-Remove-nostdlib-and-musl_root-from-musl-targets.patch"
-	"${FILESDIR}/0006-Prefer-libgcc_eh-over-libunwind-for-musl.patch"
-	"${FILESDIR}/0007-Add-target-thumbv7neon-unknown-linux-musleabihf.patch"
-	"${FILESDIR}/0008-Configure-LLVM-module-PIC-level.patch"
-	"${FILESDIR}/0009-Fix-C-aggregate-passing-ABI-on-powerpc.patch"
-	"${FILESDIR}/0010-Fix-zero-sized-aggregate-ABI-on-powerpc.patch"
-	"${FILESDIR}/0011-Link-libssp_nonshared.a-on-all-musl-targets.patch"
-	"${FILESDIR}/0012-Use-rustc-workspace-hack-for-rustbook.patch"
-	"${FILESDIR}/0013-test-failed-doctest-output-Fix-normalization.patch"
-	"${FILESDIR}/0014-test-use-extern-for-plugins-Don-t-assume-multilib.patch"
-	"${FILESDIR}/0015-test-sysroot-crates-are-unstable-Fix-test-when-rpath.patch"
-	"${FILESDIR}/0016-Ignore-broken-and-non-applicable-tests.patch"
-	"${FILESDIR}/0017-Link-stage-2-tools-dynamically-to-libstd.patch"
-	"${FILESDIR}/0018-Move-debugger-scripts-to-usr-share-rust.patch"
-	"${FILESDIR}/0019-Add-gentoo-target-specs.patch"
-	"${FILESDIR}/0030-libc-linkage.patch"
-	"${FILESDIR}/0031-typenum-pmmx.patch"
-	"${FILESDIR}/0032-libgit2-sys-abi.patch"
-	"${FILESDIR}/0040-rls-atomics.patch"
-	"${FILESDIR}/0050-llvm.patch"
-	"${FILESDIR}/0051-llvm-powerpc-plt.patch"
-	"${FILESDIR}/0052-llvm-powerpc-elfv2.patch"
-	"${FILESDIR}/0060-Update-libressl-support-3.x.patch"
+	"${FILESDIR}"/1.40.0-add-soname.patch
+	"${FILESDIR}"/1.41.1-llvm10.patch
+	"${FILESDIR}"/llvm-gcc10.patch
+	"${FILESDIR}"/0001-Don-t-pass-CFLAGS-to-the-C-compiler.patch
+	"${FILESDIR}"/0002-Fix-LLVM-build.patch
+	"${FILESDIR}"/0003-Allow-rustdoc-to-work-when-cross-compiling-on-musl.patch
+	"${FILESDIR}"/0004-Require-static-native-libraries-when-linking-static-.patch
+	"${FILESDIR}"/0005-Remove-nostdlib-and-musl_root-from-musl-targets.patch
+	"${FILESDIR}"/0006-Prefer-libgcc_eh-over-libunwind-for-musl.patch
+	"${FILESDIR}"/0007-Fix-zero-sized-aggregate-ABI-on-powerpc.patch
+	"${FILESDIR}"/0008-Link-libssp_nonshared.a-on-all-musl-targets.patch
+	"${FILESDIR}"/0009-test-failed-doctest-output-Fix-normalization.patch
+	"${FILESDIR}"/0010-test-use-extern-for-plugins-Don-t-assume-multilib.patch
+	"${FILESDIR}"/0011-test-sysroot-crates-are-unstable-Fix-test-when-rpath.patch
+	"${FILESDIR}"/0012-Ignore-broken-and-non-applicable-tests.patch
+	"${FILESDIR}"/0013-Link-stage-2-tools-dynamically-to-libstd.patch
+	"${FILESDIR}"/0014-Move-debugger-scripts-to-usr-share-rust.patch
+	"${FILESDIR}"/0015-Add-gentoo-target-specs.patch
+	"${FILESDIR}"/0030-libc-linkage.patch
+	"${FILESDIR}"/0031-typenum-pmmx.patch
+	"${FILESDIR}"/0040-rls-atomics.patch
+	"${FILESDIR}"/0050-llvm.patch
+	"${FILESDIR}"/0051-llvm-powerpc-plt.patch
+	"${FILESDIR}"/0052-llvm-powerpc-elfv2.patch
 )
 
 S="${WORKDIR}/${MY_P}-src"
@@ -130,7 +142,7 @@ pre_build_checks() {
 	CHECKREQS_DISK_BUILD="9G"
 	eshopts_push -s extglob
 	if is-flagq '-g?(gdb)?([1-9])'; then
-		CHECKREQS_DISK_BUILD="14G"
+		CHECKREQS_DISK_BUILD="15G"
 	fi
 	eshopts_pop
 	check-reqs_pkg_setup
@@ -144,28 +156,33 @@ pkg_setup() {
 	pre_build_checks
 	python-any-r1_pkg_setup
 
-	export LIBGIT2_SYS_USE_PKG_CONFIG=1
+	# use bundled for now, #707746
+	# will need dev-libs/libgit2 slotted dep if re-enabled
+	#export LIBGIT2_SYS_USE_PKG_CONFIG=1
 	export LIBSSH2_SYS_USE_PKG_CONFIG=1
 	export PKG_CONFIG_ALLOW_CROSS=1
 
 	if use system-llvm; then
 		llvm_pkg_setup
+
 		local llvm_config="$(get_llvm_prefix "$LLVM_MAX_SLOT")/bin/llvm-config"
 
 		export LLVM_LINK_SHARED=1
-		export RUSTFLAGS="$RUSTFLAGS -Lnative=$("$llvm_config" --libdir)"
+		export RUSTFLAGS="${RUSTFLAGS} -Lnative=$("${llvm_config}" --libdir)"
 	fi
 }
 
 src_prepare() {
-	default
+	if ! use system-bootstrap; then
+		local rust_stage0_root="${WORKDIR}"/rust-stage0
+		local rust_stage0="rust-${RUST_STAGE0_VERSION}-${CHOST}"
 
-	"${WORKDIR}/rust-${RUST_STAGE0_VERSION}-${CHOST}/install.sh" \
-		--destdir="${WORKDIR}/stage0" \
-		--prefix=/ \
-		--components=rust-std-$CHOST,rustc,cargo \
-		--disable-ldconfig \
-		|| die
+		"${WORKDIR}/${rust_stage0}"/install.sh --disable-ldconfig \
+			--components=rust-std-${CHOST},rustc,cargo \
+			--destdir="${rust_stage0_root}" --prefix=/ || die
+	fi
+
+	default
 }
 
 src_configure() {
@@ -185,11 +202,25 @@ src_configure() {
 		rust_targets="${rust_targets},\"wasm32-unknown-unknown\""
 	fi
 
-	for tool in clippy miri rls rustfmt; do
-		if use $tool; then
-			tools+=", \"$tool\""
-		fi
-	done
+	if use clippy; then
+		tools="\"clippy\",$tools"
+	fi
+	if use miri; then
+		tools="\"miri\",$tools"
+	fi
+	if use rls; then
+		tools="\"rls\",\"analysis\",\"src\",$tools"
+	fi
+	if use rustfmt; then
+		tools="\"rustfmt\",$tools"
+	fi
+
+	local rust_stage0_root
+	if use system-bootstrap; then
+		rust_stage0_root="$(rustc --print sysroot)"
+	else
+		rust_stage0_root="${WORKDIR}"/rust-stage0
+	fi
 
 	cat <<- EOF > "${S}"/config.toml
 		[llvm]
@@ -204,8 +235,8 @@ src_configure() {
 		build = "${CHOST}"
 		host = ["${CHOST}"]
 		target = [${rust_targets}]
-		cargo = "${WORKDIR}/stage0/bin/cargo"
-		rustc = "${WORKDIR}/stage0/bin/rustc"
+		cargo = "${rust_stage0_root}/bin/cargo"
+		rustc = "${rust_stage0_root}/bin/rustc"
 		docs = $(toml_usex doc)
 		compiler-docs = $(toml_usex doc)
 		submodules = false
@@ -214,14 +245,15 @@ src_configure() {
 		vendor = true
 		extended = true
 		tools = [${tools}]
-		verbose = 0
+		verbose = 2
 		sanitizers = false
 		profiler = false
 		cargo-native-static = false
 		[install]
 		prefix = "${EPREFIX}/usr"
-		docdir = "share/doc/${P}"
-		mandir = "share/${P}/man"
+		libdir = "lib"
+		docdir = "share/doc/${PF}"
+		mandir = "share/man"
 		[rust]
 		optimize = true
 		debug = $(toml_usex debug)
@@ -229,9 +261,9 @@ src_configure() {
 		debuginfo-level-rustc = 0
 		backtrace = true
 		incremental = false
-		parallel-compiler = false
 		default-linker = "$(tc-getCC)"
-		channel = "stable"
+		parallel-compiler = $(toml_usex parallel-compiler)
+		channel = "$(usex nightly nightly stable)"
 		rpath = false
 		verbose-tests = true
 		optimize-tests = $(toml_usex !debug)
@@ -249,9 +281,11 @@ src_configure() {
 		ar = "$(tc-getAR)"
 		crt-static = false
 	EOF
-	use system-llvm && cat <<- EOF >> "${S}"/config.toml
-		llvm-config = "$(get_llvm_prefix "$LLVM_MAX_SLOT")/bin/llvm-config"
-	EOF
+	if use system-llvm; then
+		cat <<- EOF >> "${S}"/config.toml
+			llvm-config = "$(get_llvm_prefix "${LLVM_MAX_SLOT}")/bin/llvm-config"
+		EOF
+	fi
 
 	for v in $(multilib_get_enabled_abi_pairs); do
 		rust_target=$(rust_abi $(get_abi_CHOST ${v##*.}) | sed s/gnu/musl/)
@@ -295,11 +329,11 @@ src_configure() {
 
 src_compile() {
 	RUST_BACKTRACE=1 \
-	"${EPYTHON}" x.py build --config="${S}"/config.toml -j$(makeopts_jobs) || die
+	"${EPYTHON}" ./x.py build -vv --config="${S}"/config.toml -j$(makeopts_jobs) || die
 }
 
 src_test() {
-	"${EPYTHON}" x.py test -j$(makeopts_jobs) --no-doc --no-fail-fast \
+	"${EPYTHON}" ./x.py test -vv --config="${S}"/config.toml -j$(makeopts_jobs) --no-doc --no-fail-fast \
 		src/test/codegen \
 		src/test/codegen-units \
 		src/test/compile-fail \
@@ -318,7 +352,12 @@ src_test() {
 }
 
 src_install() {
-	env DESTDIR="${D}" "${EPYTHON}" x.py install || die
+	env DESTDIR="${D}" "${EPYTHON}" ./x.py install -vv --config="${S}"/config.toml || die
+
+	# bug #689562, #689160
+	rm "${D}/etc/bash_completion.d/cargo" || die
+	rmdir "${D}"/etc{/bash_completion.d,} || die
+	dobashcomp build/tmp/dist/cargo-image/etc/bash_completion.d/cargo
 
 	mv "${ED}/usr/bin/cargo" "${ED}/usr/bin/cargo-${PV}" || die
 	mv "${ED}/usr/bin/rustc" "${ED}/usr/bin/rustc-${PV}" || die
@@ -327,7 +366,6 @@ src_install() {
 	mv "${ED}/usr/bin/rust-gdbgui" "${ED}/usr/bin/rust-gdbgui-${PV}" || die
 	mv "${ED}/usr/bin/rust-lldb" "${ED}/usr/bin/rust-lldb-${PV}" || die
 
-	rm "${ED}/usr/lib"/*.so || die
 	rm "${ED}/usr/lib/rustlib/components" || die
 	rm "${ED}/usr/lib/rustlib/install.log" || die
 	rm "${ED}/usr/lib/rustlib"/manifest-* || die
@@ -349,24 +387,17 @@ src_install() {
 		mv "${ED}/usr/bin/rustfmt" "${ED}/usr/bin/rustfmt-${PV}" || die
 	fi
 
-	if use doc; then
-		dodir "/usr/share/doc/${P}"
-		mv "${ED}/usr/share/doc/rust"/* "${ED}/usr/share/doc/${P}" || die
-		rmdir "${ED}/usr/share/doc/rust" || die
+	# Move public shared libs to abi specific libdir
+	# Private and target specific libs MUST stay in /usr/lib/rustlib/${rust_target}/lib
+	if [[ $(get_libdir) != lib ]]; then
+		dodir /usr/$(get_libdir)
+		mv "${ED}/usr/lib"/*.so "${ED}/usr/$(get_libdir)/" || die
 	fi
 
 	dodoc COPYRIGHT
 	rm "${ED}/usr/share/doc/${P}"/*.old || die
 	rm "${ED}/usr/share/doc/${P}/LICENSE-APACHE" || die
 	rm "${ED}/usr/share/doc/${P}/LICENSE-MIT" || die
-
-	docompress "/usr/share/${P}/man"
-
-	cat <<-EOF > "${T}"/50${P}
-		LDPATH="${EPREFIX}/usr/lib/rustlib/${CHOST}/lib"
-		MANPATH="${EPREFIX}/usr/share/${P}/man"
-	EOF
-	doenvd "${T}"/50${P}
 
 	# note: eselect-rust adds EROOT to all paths below
 	cat <<-EOF > "${T}/provider-${P}"
@@ -390,7 +421,7 @@ src_install() {
 		echo /usr/bin/cargo-fmt >> "${T}/provider-${P}"
 		echo /usr/bin/rustfmt >> "${T}/provider-${P}"
 	fi
-	dodir /etc/env.d/rust
+
 	insinto /etc/env.d/rust
 	doins "${T}/provider-${P}"
 }
@@ -405,16 +436,12 @@ pkg_postinst() {
 	ewarn "This might have resulted in a dangling symlink for /usr/bin/cargo on some"
 	ewarn "systems. This can be resolved by calling 'sudo eselect rust set ${P}'."
 
-	if has_version app-editors/emacs || has_version app-editors/emacs-vcs; then
+	if has_version app-editors/emacs; then
 		elog "install app-emacs/rust-mode to get emacs support for rust."
 	fi
 
 	if has_version app-editors/gvim || has_version app-editors/vim; then
 		elog "install app-vim/rust-vim to get vim support for rust."
-	fi
-
-	if has_version 'app-shells/zsh'; then
-		elog "install app-shells/rust-zshcomp to get zsh completion for rust."
 	fi
 }
